+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## Unreleased
+
+### Changed
+
+- Remove `instructor` package.
+
 ## 0.1.0
 
 ### Added